+++ src/widgets/OverlayWindow.cpp
@@ -96,7 +96,7 @@ namespace chatterino {
 using namespace std::chrono_literals;
 
 OverlayWindow::OverlayWindow(IndirectChannel channel,
-                             const QList<QUuid> &filterIDs)
+                             const QList<QUuid> &filterIDs, bool filtersAnyOf)
     : BaseWindow({
           BaseWindow::Frameless,
           BaseWindow::TopMost,
@@ -130,7 +130,7 @@ OverlayWindow::OverlayWindow(IndirectChannel channel,
     });
 
     this->channelView_.installEventFilter(this);
-    this->channelView_.setFilters(filterIDs);
+    this->channelView_.setFilters(filterIDs, filtersAnyOf);
     this->channelView_.setChannel(this->channel_.get());
     this->channelView_.setIsOverlay(true);  // use overlay colors
     this->channelView_.setAttribute(Qt::WA_TranslucentBackground);